Branch point of periodic orbit

Branch point of periodic orbit

Predictor

Which method to use?

Normal form based on Poincaré return map

References

At a Branch point (BP) of a periodic orbit $\gamma$ (with period $T$) for the Cauchy problem at parameter value $p_0$:

\[\frac{du}{dt}=F(u,p),\tag{E}\]

the eigenvalues (Floquet coefficients) of the monodromy operator $\mathcal M=Y(T)$ solution to

\[\frac{dY}{dt}=A(t)Y(t), Y(0)=I_n\]

contain the eigenvalue $1$ with algebraic multiplicity 2. There are two ways to compute the normal form of this bifurcation:

using the Poincaré return map [Kuznetsov],

using the method of [Iooss], see also [Kuz2].

You can obtain the normal form of a BP bifurcation using

pd = get_normal_form(br, ind; prm = Val(false))

where prm indicates whether you want to use the method based on Poincaré return map (PRM) or the one based on Iooss method.

BP nf computation

The method of [Iooss] is not implemented.

Predictor

The predictor for a non trivial guess at distance $\delta p$ from the bifurcation point is provided by the method:

BifurcationKit.predictor — Method

predictor(
    nf::BifurcationKit.BranchPointPO{<:Shooting},
    δp,
    ampfactor;
    override
) -> NamedTuple{(:orbitguess, :pnew, :prob, :ampfactor, :po), <:Tuple{Any, Any, Shooting, Any, Any}}

Compute the predictor for the simple branch point of periodic orbit.

If override = true, then the predictor is simply x0 .+ ampfactor .* e for the parameter p0 + δp and where e is a bifurcating eigenvector.

Which method to use?

Depending on the method used for computing the periodic orbits, you have several possibilities:

For shooting, you can only the PRM method. Shooting is the preferred way for large scale systems. Note that the PRM method is not very precise numerically.

For collocation, you can use PRM and Iooss methods. Note that the Iooss method is the most precise. This is not yet implemented.

For Trapezoid method, the normal form is not yet implemented.

Normal form based on Poincaré return map

Given a transversal section $\Sigma$ to $\gamma$ at $\gamma(0)$, the Poincaré return map $\mathcal P$ associates to each point $x\in\Sigma$ close to $\gamma(0)$ the first point $\mathcal P(x,p)\in\Sigma$ where the orbit of (E) with initial condition $x$ intersects again $\Sigma$. Hence, the discrete map $x_{n+1}=\mathcal P(x_n,p)$ has normal form

\[x_{n+1} = x_n + a_{10}\delta p + a_{11}\delta p x + a_{02}x^2 + a_{03}x^3\]

Large scale problems

The computation of the normal form is not optimized for Matrix-Free problems (e.g. Monodromy) yet.

References

Kuznetsov

Yu. A. Kuznetsov, "Elements of Applied Bifurcation Theory", 2nd ed., 1998.

Kuz2

Kuznetsov et al., “Numerical Periodic Normalization for Codim 1 Bifurcations of Limit Cycles.”

Iooss

Iooss, "Global Characterization of the Normal Form for a Vector Field near a Closed Orbit.", 1988

predictor( nf::BifurcationKit.BranchPointPO{<:Shooting}, δp, ampfactor; override ) -> NamedTuple{(:orbitguess, :pnew, :prob, :ampfactor, :po), <:Tuple{Any, Any, Shooting, Any, Any}}
Compute the predictor for the simple branch point of periodic orbit.

predictor( gh::BifurcationKit.Bautin, ::Val{:FoldPeriodicOrbitCont}, ϵ; verbose, ampfactor ) -> NamedTuple{(:orbit, :ω, :params, :x0), <:Tuple{BifurcationKit.var"#786#789"{BifurcationKit.var"#FoldPO#788"{ϵ, q0, x0, h₂₀₀₀}} where {ϵ, q0, x0, h₂₀₀₀}, Any, Any, BifurcationKit.var"#787#790"}}
Compute the predictor for the curve of Folds of periodic orbits near the Bautin bifurcation point.

predictor( bt::BifurcationKit.BogdanovTakens, ::Val{:FoldCurve}, ds; verbose, ampfactor ) -> NamedTuple{(:fold, :EigenVec, :EigenVecAd, :x0), <:Tuple{BifurcationKit.var"#FoldCurve#741", BifurcationKit.var"#737#742"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#738#743"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#739#744"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#getx#740"{_A}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf, _A}}}
Compute the predictor for the Fold curve near the Bogdanov-Takens point.

predictor( bt::BifurcationKit.BogdanovTakens, ::Val{:HomoclinicCurve}, ds; verbose, ampfactor ) -> NamedTuple{(:α, :orbit), <:Tuple{BifurcationKit.var"#α#746", BifurcationKit.var"#xLP#752"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#u2#751"{a1, d, b, a}, BifurcationKit.var"#v1#750"{BifurcationKit.var"#v0#748", _A, _B}, BifurcationKit.var"#u1#749", BifurcationKit.var"#v0#748", BifurcationKit.var"#u0#747"} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf, a1, d, b, a, _A, _B}}}
Compute the predictor for the curve of homoclinic orbits near the Bogdanov-Takens point.

predictor( bt::BifurcationKit.BogdanovTakens, ::Val{:HopfCurve}, ds; verbose, ampfactor ) -> NamedTuple{(:hopf, :ω, :EigenVec, :EigenVecAd, :x0), <:Tuple{BifurcationKit.var"#726#733"{BifurcationKit.var"#HopfCurve#730"{par0, K2, K11, K10, b, a}} where {par0, K2, K11, K10, b, a}, BifurcationKit.var"#727#734"{BifurcationKit.var"#HopfCurve#730"{par0, K2, K11, K10, b, a}} where {par0, K2, K11, K10, b, a}, BifurcationKit.var"#EigenVec#731"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#getx#729"{_A}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf, _A}, BifurcationKit.var"#EigenVecAd#732"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#getx#729"{_A}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf, _A}, BifurcationKit.var"#728#735"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#getx#729"{_A}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf, _A}}}
Compute the predictor for the Hopf curve near the Bogdanov-Takens point.

predictor( hh::BifurcationKit.HopfHopf, ::Val{:HopfCurve}, ds; verbose, ampfactor ) -> NamedTuple{(:hopf, :ω, :EigenVec, :EigenVecAd, :x0), <:Tuple{BifurcationKit.var"#872#878"{BifurcationKit.var"#HopfCurve#875"{_A, _B}} where {_A, _B}, BifurcationKit.var"#873#879"{BifurcationKit.var"#HopfCurve#875"{_A, _B}} where {_A, _B}, BifurcationKit.var"#EigenVec#876"{BifurcationKit.HopfHopf{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#EigenVecAd#877"{BifurcationKit.HopfHopf{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#874#880"}}
Compute the predictor for the Hopf curve near the Hopf-Hopf bifurcation point.

predictor( hh::BifurcationKit.HopfHopf, ::Val{:NS}, ϵ; verbose, ampfactor ) -> NamedTuple{(:ns1, :ns2, :params1, :params2, :ω11, :ω12, :ω21, :ω22, :T1, :T2, :k1, :k2), <:Tuple{BifurcationKit.var"#882#886"{BifurcationKit.var"#NS1#884"{ϵ, q1, x1, h₂₀₀₀}} where {ϵ, q1, x1, h₂₀₀₀}, BifurcationKit.var"#883#887"{BifurcationKit.var"#NS2#885"{ϵ, q2, x2, h₀₀₂₀}} where {ϵ, q2, x2, h₀₀₂₀}, Vararg{Any, 10}}}
Compute the predictor for the curve of Neimark-Sacker points near the Hopf-Hopf bifurcation point.

struct Shooting{Tf<:BifurcationKit.AbstractFlow, Tjac<:BifurcationKit.AbstractJacobianType, Ts, Tsection, Tpar, Tlens} <: BifurcationKit.AbstractShootingDiscretization
Create a problem to implement the Simple / Parallel Multiple Standard Shooting method to locate periodic orbits / BVP.

predictor( nf::BifurcationKit.PeriodDoublingPO{<:Shooting}, δp, ampfactor; override ) -> NamedTuple{(:orbitguess, :pnew, :prob, :ampfactor, :δp, :po), <:NTuple{6, Any}}
Compute the predictor for the period-doubling bifurcation of periodic orbit.

predictor( zh::BifurcationKit.ZeroHopf, ::Val{:FoldCurve}, ds; verbose, ampfactor ) -> NamedTuple{(:fold, :λ0, :EigenVec, :EigenVecAd, :x0), <:Tuple{BifurcationKit.var"#833#839"{BifurcationKit.var"#FoldCurve#836"{_A, _B}} where {_A, _B}, BifurcationKit.var"#834#840"{BifurcationKit.var"#FoldCurve#836"{_A, _B}} where {_A, _B}, BifurcationKit.var"#EigenVec#837"{BifurcationKit.ZeroHopf{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#EigenVecAd#838"{BifurcationKit.ZeroHopf{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#835#841"}}
Compute the predictor for the curve of Fold bifurcations near the Zero-Hopf bifurcation point.

predictor( zh::BifurcationKit.ZeroHopf, ::Val{:HopfCurve}, ds; verbose, ampfactor ) -> NamedTuple{(:hopf, :ω, :EigenVec, :EigenVecAd, :x0), <:Tuple{BifurcationKit.var"#823#829"{BifurcationKit.var"#HopfCurve#826"{_A, _B}} where {_A, _B}, BifurcationKit.var"#824#830"{BifurcationKit.var"#HopfCurve#826"{_A, _B}} where {_A, _B}, BifurcationKit.var"#EigenVec#827"{BifurcationKit.ZeroHopf{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#EigenVecAd#828"{BifurcationKit.ZeroHopf{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#825#831"}}
Compute the predictor for the curve of Hopf bifurcations near the Zero-Hopf bifurcation point.

predictor( zh::BifurcationKit.ZeroHopf, ::Val{:NS}, ϵ; verbose, ampfactor ) -> NamedTuple{(:orbit, :hasNS, :params, :T, :k), <:Tuple{BifurcationKit.var"#843#845"{BifurcationKit.var"#NS#844"{_A, _B, _C}} where {_A, _B, _C}, Vararg{Any, 4}}}
Compute the predictor for the curve of Neimark-Sacker bifurcations near the Zero-Hopf bifurcation point.