Bogdanov-Takens refinement

Bogdanov-Takens refinement

Method

Setting the jacobian

Example

Newton refinement

Advanced use

References

In this page, we explain how to perform precise localisation of Bogdanov-Takens (BT) points. This is an unusual feature of numerical continuation libraries. We chose to implement it because the localisation of the BT points on the Hopf bifurcation curves is rather imprecise.

Method

The continuation of BT bifurcation points is based on a Minimally Augmented[Govaerts],[Blank],[Bindel] formulation which is an efficient way to detect singularities. The continuation of BT points is based on the formulation

\[G(u,p) = (F(u,p), g_1(u,p), g_2(u,p))\in\mathbb R^{n+2}\quad\quad (F_{bt})\]

where the test functions $g_1,g_2$ are solutions of

\[\left[\begin{array}{cc} dF(u,p) & a \\ b^{\top} & 0 \end{array}\right]\left[\begin{array}{c} v_1 \\ \sigma_1(u,p) \end{array}\right]=\left[\begin{array}{c}0_{n} \\1\end{array}\right]\quad\quad (M_{bt})\]

and

\[\left[\begin{array}{cc} dF(u,p) & a \\ b^{\top} & 0 \end{array}\right]\left[\begin{array}{c} v_2 \\ \sigma_2(u,p) \end{array}\right]=\left[\begin{array}{c}b_1 \\0\end{array}\right]\quad\quad (M_{bt})\]

and where $a,b$ are chosen in order to have a non-singular matrix $(M_{bt})$. More precisely, $b$ (resp. $a$) should be close to a null vector of dF(u,p) (resp. dF(u,p)').

note that there are very simplified calls for this, see Newton refinement below. In particular, you don't need to set up the Minimally Augmented problem yourself. This is done in the background.

Linear Method

You can pass the bordered linear solver to solve $(M_{bt})$ using the option bdlinsolver (see below). Note that the choice bdlinsolver = BorderingBLS() can lead to singular systems. Indeed, in this case, $(M_{bt})$ is solved by inverting dF(u,p) which is singular at Fold points.

Setting the jacobian

In order to apply the newton algorithm to $F_{bt}$, one needs to invert the jacobian. This is not completely trivial as one must compute this jacobian and then invert it. You can select the following jacobians for your computations (see below):

jacobian_ma = AutoDiff(): [Default] automatic differentiation is applied to $F_{bt}$ and the matrix is then inverted using the provided linear solver. In particular, the jacobian is formed. This is very well suited for small dimensions (say < 100)

jacobian_ma = MinAug(): a specific procedure for evaluating the jacobian $F_{bt}$ and inverting it (without forming the jacobian!) is used. This is well suited for large dimensions.

Example

using Revise, BifurcationKitFbt(x, p) = [x[2], p.β1 + p.β2 * x[2] + p.a * x[1]^2 + p.b * x[1] * x[2]]par = (β1 = 0.01, β2 = -0.3, a = -1., b = 1.)prob  = BifurcationProblem(Fbt, [0.01, 0.01], par, (@optic _.β1))opts_br = ContinuationPar(p_max = 0.5, p_min = -0.5, detect_bifurcation = 3, nev = 2)br = continuation(prob, PALC(), opts_br; bothside = true)# compute branch of Hopf pointshopf_codim2 = continuation(br, 3, (@optic _.β2), ContinuationPar(opts_br, max_steps = 40) ;	detect_codim2_bifurcation = 2,	update_minaug_every_step = 1,	bothside = true,	)# refine BT pointsolbt = BifurcationKit.newton_bt(hopf_codim2, 2)solbt.u

Bogdanov-Takens bifurcation point at (:β1, :β2) ≈ (0.0, 0.0).
Normal form (B, p1 + p2⋅B + b⋅A⋅B + a⋅A²)
Normal form coefficients:
 a = missing
 b = missing

You can call various predictors:
 - predictor(::BogdanovTakens, ::Val{:HopfCurve}, ds)
 - predictor(::BogdanovTakens, ::Val{:FoldCurve}, ds)
 - predictor(::BogdanovTakens, ::Val{:HomoclinicCurve}, ds)

Newton refinement

Once a Bogdanov-Takens point has been detected after a call to br = continuation(...), it can be refined using newton iterations. Let us say that ind_bif is the index in br.specialpoint of a Bogdanov-Takens point. This guess can be refined as follows:

outfold = newton(br::AbstractBranchResult, ind_bif::Int;  	normN = norm,	options = br.contparams.newton_options,	start_with_eigen = false,	lens2 = (@optic _), kwargs...)

For the options parameters, we refer to Krylov-Newton algorithm. Note that you can pass a bordered linear solver or a specific jacobian type (jacobian_ma) through the option of newton_bt.

It is important to note that for improved performances, a function implementing the expression of the hessian should be provided. This is by far the fastest. BifurcationProblem provides it by default using AD though.

Advanced use

Here, we expose the solvers that are used to perform newton refinement. This is useful in case it is too involved to expose the linear solver options.

BifurcationKit.newton_bt — Function

newton_bt(
    prob::BifurcationKit.AbstractBifurcationProblem,
    btpointguess::BorderedArray,
    par,
    lens2::Union{typeof(identity), IndexLens, PropertyLens, ComposedFunction},
    eigenvec,
    eigenvec_ad,
    options::NewtonPar;
    normN,
    jacobian_ma,
    usehessian,
    bdlinsolver,
    bdlinsolver_adjoint,
    bdlinsolver_block,
    kwargs...
) -> Any

This function turns an initial guess for a BT point into a solution to the BT problem based on a Minimally Augmented formulation. The arguments are as follows

prob::AbstractBifurcationProblem

btpointguess initial guess (x0, p0) for the BT point. It should be a BorderedArray as returned by the function bt_point

par parameters used for the vector field

eigenvec guess for the 0 eigenvector

eigenvec_ad guess for the 0 adjoint eigenvector

options::NewtonPar options for the Newton-Krylov algorithm, see NewtonPar.

Optional arguments:

normN = norm

bdlinsolver bordered linear solver for the constraint equation

jacobian_ma::AbstractJacobianType = AutoDiff() how the linear system (for newton) is solved. Can be (AutoDiff(), FiniteDifferences(), MinAug())

kwargs keywords arguments to be passed to the regular Newton-Krylov solver

Simplified call

Simplified call to refine an initial guess for a BT point. More precisely, the call is as follows

newton(br::AbstractBranchResult, ind_bt::Int; options = br.contparams.newton_options, kwargs...)

The parameters / options are as usual except that you have to pass the branch br from the result of a call to continuation with detection of bifurcations enabled and index is the index of bifurcation point in br you want to refine. You can pass newton parameters different from the ones stored in br by using the argument options.

Jacobian transpose

The adjoint of the jacobian J is computed internally when Jᵗ = nothing by using transpose(J) which works fine when J is an AbstractArray. In this case, do not pass the jacobian adjoint like Jᵗ = (x, p) -> transpose(d_xF(x, p)) otherwise the jacobian will be computed twice!

ODE problems

For ODE problems, it is more efficient to pass the option jacobian_ma = :autodiff

newton_bt(
    br::BifurcationKit.AbstractResult{Tkind, Tprob<:Union{BifurcationKit.FoldMAProblem, BifurcationKit.HopfMAProblem}},
    ind_bt::Int64;
    probvf,
    normN,
    options,
    nev,
    start_with_eigen,
    a,
    b,
    bdlinsolver,
    bdlinsolver_adjoint,
    kwargs...
) -> Any

This function turns an initial guess for a Bogdanov-Takens point into a solution to the Bogdanov-Takens problem based on a Minimally Augmented formulation.

Arguments

br results returned after a call to continuation

ind_bif bifurcation index in br

Optional arguments:

options::NewtonPar, default value br.contparams.newton_options

normN = norm

bdlinsolver bordered linear solver for the constraint equation

start_with_eigen = false whether to start the Minimally Augmented problem with information from eigen elements. If start_with_eigen = false, then:

a::Nothing estimate of null vector. If nothing is passed, a random vector is used. In case you do not rely on AbstractArray, you should probably pass this.

b::Nothing estimate of second null vector. If nothing is passed, a random vector is used. In case you do not rely on AbstractArray, you should probably pass this.

kwargs keywords arguments to be passed to the regular Newton-Krylov solver

ODE problems

For ODE problems, it is more efficient to pass the option jacobian = :autodiff

start_with_eigen

For ODE problems, it is more efficient to pass the option start_with_eigen = true

References

Govaerts

Govaerts, Willy J. F. Numerical Methods for Bifurcations of Dynamical Equilibria. Philadelphia, Pa: Society for Industrial and Applied Mathematics, 2000.

Blank

Blank, H. J. de, Yu. A. Kuznetsov, M. J. Pekkér, and D. W. M. Veldman. “Degenerate Bogdanov–Takens Bifurcations in a One-Dimensional Transport Model of a Fusion Plasma.” Physica D: Nonlinear Phenomena 331 (September 15, 2016): 13–26. https://doi.org/10.1016/j.physd.2016.05.008.

Bindel

Bindel, D., M. Friedman, W. Govaerts, J. Hughes, and Yu.A. Kuznetsov. “Numerical Computation of Bifurcations in Large Equilibrium Systems in Matlab.” Journal of Computational and Applied Mathematics 261 (May 2014): 232–48. https://doi.org/10.1016/j.cam.2013.10.034.

newton_bt( br::BifurcationKit.AbstractResult{Tkind, Tprob<:Union{BifurcationKit.FoldMAProblem, BifurcationKit.HopfMAProblem}}, ind_bt::Int64; probvf, normN, options, nev, start_with_eigen, a, b, bdlinsolver, bdlinsolver_adjoint, kwargs... ) -> Any
This function turns an initial guess for a Bogdanov-Takens point into a solution to the Bogdanov-Takens problem based on a Minimally Augmented formulation.

mutable struct BorderedArray{𝒯1, 𝒯2}
This defines an "array" (although not <: AbstractArray) to hold two arrays or an array and a scalar.

continuation( br::BifurcationKit.AbstractBranchResult, ind_bif, lens2::Union{typeof(identity), IndexLens, PropertyLens, ComposedFunction}; ... ) -> Any continuation( br::BifurcationKit.AbstractBranchResult, ind_bif, lens2::Union{typeof(identity), IndexLens, PropertyLens, ComposedFunction}, options_cont::ContinuationPar; prob, start_with_eigen, detect_codim2_bifurcation, update_minaug_every_step, kwargs... ) -> Any
Codimension 2 continuation of Fold / Hopf points.

struct BifurcationProblem{Tvf, Tu, Tp, Tl<:Union{typeof(identity), IndexLens, PropertyLens, ComposedFunction}, Tplot, Trec, Tgets, Tupdate} <: BifurcationKit.AbstractAllJetBifProblem
Structure to hold a bifurcation problem.

struct ContinuationPar{T, S<:BifurcationKit.AbstractLinearSolver, E<:AbstractEigenSolver}
Returns a variable containing the parameters to affect the continuation algorithm used to solve F(x, p) = 0.

struct NewtonPar{T, L<:BifurcationKit.AbstractLinearSolver, E<:AbstractEigenSolver}
Returns a variable containing parameters to affect the newton algorithm when solving F(x) = 0.

struct PALC{Ttang<:BifurcationKit.AbstractTangentComputation, Tbls<:BifurcationKit.AbstractLinearSolver, T, Tdot} <: BifurcationKit.AbstractContinuationAlgorithm
Pseudo-arclength continuation algorithm.

continuation( br::BifurcationKit.AbstractResult{Tkind<:Union{BifurcationKit.BoundaryValueProblemCont, BifurcationKit.EquilibriumCont}, Tprob}, ind_bif::Int64; ... ) -> Any continuation( br::BifurcationKit.AbstractResult{Tkind<:Union{BifurcationKit.BoundaryValueProblemCont, BifurcationKit.EquilibriumCont}, Tprob}, ind_bif::Int64, options_cont::ContinuationPar; ... ) -> Branch continuation( br::BifurcationKit.AbstractResult{Tkind<:Union{BifurcationKit.BoundaryValueProblemCont, BifurcationKit.EquilibriumCont}, Tprob}, ind_bif::Int64, options_cont::ContinuationPar, Teigvec::Type{𝒯eigvec}; alg, δp, ampfactor, use_normal_form, bls, bls_block, nev, scaleζ, autodiff, usedeflation, verbosedeflation, max_iter_deflation, perturb, plot_solution, tol_fold, kwargs_deflated_newton, kwargs... ) -> Any
Automatic branch switching at branch points based on a computation of the normal form.

continuation(::TWModel, ::Any, ::BifurcationKit.AbstractContinuationAlgorithm, ::ContinuationPar)
Specific continuation routine for wave problems.

continuation( prob::BifurcationKit.AbstractBifurcationProblem, alg::BifurcationKit.AbstractContinuationAlgorithm, contparams::ContinuationPar; linear_algo, bothside, kwargs... ) -> BifurcationKit.DCResult{Tprob, Tbr, Tit, Tsol, Talg} where {Tprob<:BifurcationKit.BVP.BVPBifProblem, Tbr<:(Vector), Tit<:(ContIterable{BifurcationKit.BoundaryValueProblemCont, Tprob, Talg, T, _A, _B, typeof(LinearAlgebra.norm), typeof(BifurcationKit.finalise_default), typeof(BifurcationKit.cb_default), Nothing} where {Tprob<:BifurcationKit.BVP.BVPBifProblem, Talg<:BifurcationKit.AbstractContinuationAlgorithm, T<:Real, _A, _B}), Tsol<:(Vector), Talg<:DefCont} continuation( prob::BifurcationKit.AbstractBifurcationProblem, algdc::DefCont, contParams::ContinuationPar; verbosity, plot, linear_algo, dot_palc, callback_newton, filename, normC, kwcont... ) -> BifurcationKit.DCResult{Tprob, Tbr, Tit, Tsol, Talg} where {Tprob<:BifurcationKit.BVP.BVPBifProblem, Tbr<:(Vector), Tit<:(ContIterable{BifurcationKit.BoundaryValueProblemCont, Tprob, Talg, T, _A, _B, typeof(LinearAlgebra.norm), typeof(BifurcationKit.finalise_default), typeof(BifurcationKit.cb_default), Nothing} where {Tprob<:BifurcationKit.BVP.BVPBifProblem, Talg<:BifurcationKit.AbstractContinuationAlgorithm, T<:Real, _A, _B}), Tsol<:(Vector), Talg<:DefCont} continuation( br::BifurcationKit.AbstractBranchResult, ind_bif, lens2::Union{typeof(identity), IndexLens, PropertyLens, ComposedFunction}; ... ) -> Any continuation( br::BifurcationKit.AbstractBranchResult, ind_bif, lens2::Union{typeof(identity), IndexLens, PropertyLens, ComposedFunction}, options_cont::ContinuationPar; prob, start_with_eigen, detect_codim2_bifurcation, update_minaug_every_step, kwargs... ) -> Any continuation( br::BifurcationKit.AbstractResult{Tkind<:Union{BifurcationKit.BoundaryValueProblemCont, BifurcationKit.EquilibriumCont}, Tprob}, ind_bif::Int64; ... ) -> Any continuation( br::BifurcationKit.AbstractResult{Tkind<:Union{BifurcationKit.BoundaryValueProblemCont, BifurcationKit.EquilibriumCont}, Tprob}, ind_bif::Int64, options_cont::ContinuationPar; ... ) -> Branch continuation( br::BifurcationKit.AbstractResult{Tkind<:Union{BifurcationKit.BoundaryValueProblemCont, BifurcationKit.EquilibriumCont}, Tprob}, ind_bif::Int64, options_cont::ContinuationPar, Teigvec::Type{𝒯eigvec}; alg, δp, ampfactor, use_normal_form, bls, bls_block, nev, scaleζ, autodiff, usedeflation, verbosedeflation, max_iter_deflation, perturb, plot_solution, tol_fold, kwargs_deflated_newton, kwargs... ) -> Any continuation( br::BifurcationKit.AbstractResult{BifurcationKit.PeriodicOrbitCont, Tprob<:BifurcationKit.AbstractWrapperPeriodicOrbitProblem}, ind_bif::Int64, _contParams::ContinuationPar; alg, δp, ampfactor, usedeflation, linear_algo, detailed, prm, use_normal_form, autodiff_nf, kwargs... ) -> Branch
Compute the continuation curve associated to the functional F which is stored in the bifurcation problem prob.

continuation( br::BifurcationKit.AbstractBranchResult, ind_bif, lens2::Union{typeof(identity), IndexLens, PropertyLens, ComposedFunction}; ... ) -> Any continuation( br::BifurcationKit.AbstractBranchResult, ind_bif, lens2::Union{typeof(identity), IndexLens, PropertyLens, ComposedFunction}, options_cont::ContinuationPar; prob, start_with_eigen, detect_codim2_bifurcation, update_minaug_every_step, kwargs... ) -> Any continuation( br::BifurcationKit.AbstractResult{Tkind<:Union{BifurcationKit.BoundaryValueProblemCont, BifurcationKit.EquilibriumCont}, Tprob}, ind_bif::Int64; ... ) -> Any continuation( br::BifurcationKit.AbstractResult{Tkind<:Union{BifurcationKit.BoundaryValueProblemCont, BifurcationKit.EquilibriumCont}, Tprob}, ind_bif::Int64, options_cont::ContinuationPar; ... ) -> Branch continuation( br::BifurcationKit.AbstractResult{Tkind<:Union{BifurcationKit.BoundaryValueProblemCont, BifurcationKit.EquilibriumCont}, Tprob}, ind_bif::Int64, options_cont::ContinuationPar, Teigvec::Type{𝒯eigvec}; alg, δp, ampfactor, use_normal_form, bls, bls_block, nev, scaleζ, autodiff, usedeflation, verbosedeflation, max_iter_deflation, perturb, plot_solution, tol_fold, kwargs_deflated_newton, kwargs... ) -> Any continuation( disc::BifurcationKit.AbstractBoundaryValueDiscretization, orbitguess, alg::BifurcationKit.AbstractContinuationAlgorithm, _contParams::ContinuationPar; linear_algo, kwargs... ) -> Any continuation( br::BifurcationKit.AbstractBranchResult, ind_bif::Int64, _contParams::ContinuationPar, disc::BifurcationKit.AbstractBoundaryValueDiscretization; bif_prob, detailed, use_normal_form, nev, kwargs... ) -> Branch continuation( trap::Trapeze, orbitguess, alg::BifurcationKit.AbstractContinuationAlgorithm, _contParams::ContinuationPar; 1003, record_from_solution, linear_algo, kwargs... ) -> Any
Codimension 2 continuation of Fold / Hopf points.

newton( br::BifurcationKit.AbstractBranchResult, ind_bif::Int64; normN, options, start_with_eigen, lens2, kwargs... ) -> Any
This function turns an initial guess for a Fold / Hopf point into a solution to the Fold / Hopf problem based on a Minimally Augmented formulation.

continuation( trap::Trapeze, orbitguess, alg::BifurcationKit.AbstractContinuationAlgorithm, _contParams::ContinuationPar; 1003, record_from_solution, linear_algo, kwargs... ) -> Any
Convenience wrapper around continuation_po to continue a branch of periodic orbits computed with the Trapeze finite-difference functional.