Period-doubling point

Period-doubling point

Which method to use?

Predictor

Normal form based on Poincaré return map

Normal form based on Iooss method

See also

References

At a period-doubling (PD) bifurcation of a periodic orbit $\gamma$ (with period $T$) for parameter value $p_0$ for the Cauchy problem

\[\frac{du}{dt}=F(u,p),\tag{E}\]

the eigenvalues (Floquet coefficients) of the monodromy operator $\mathcal M=Y(T)$ solution to

\[\frac{dY}{dt}=A(t)Y(t), Y(0)=I_n\]

contain the simple eigenvalue $\mu=-1$.

There are two ways to compute the normal form of this bifurcation:

using the Poincaré return map [Kuznetsov],

using the method of [Iooss], see also [Kuz2].

You can obtain the normal form of a PD bifurcation using

pd = get_normal_form(br, ind; prm = Val(false))

where prm indicates whether you want to use the method based on Poincaré return map (Val(true)) or the one based on Iooss method (Val(false)). The call returns a point of type PeriodDoublingPO from which you can access

pd.nf.p the parameter value at the bifurcation point,

pd.T the period of the periodic orbit (the period of the bifurcated orbit is $2T$),

pd.ζ, pd.ζ★ the right / left eigenvectors,

pd.nf.nf the coefficients of the normal form.

You can pass verbose = true to get more information during the computation and, for the periodic orbits based on orthogonal collocation, detailed = Val(false) to skip the (costly) computation of the normal form.

Which method to use?

Depending on the method used for computing the periodic orbits, you have several possibilities:

For shooting, you can only use the PRM method. Shooting is the preferred way for large scale systems. Note that the PRM method is not very precise numerically.

For collocation, you can use PRM and Iooss methods. The Iooss method (prm = Val(false), the default) is the most precise.

For Trapezoid method, PD normal form is not yet implemented.

Branch switching

Contrary to the case of Neimark-Sacker point, automatic branch switching (continuation(br, ind)) is available from a PD point: the bifurcating object is again a periodic orbit (with period close to $2T$). An example is provided in Period doubling in Lur'e problem (PD aBS).

Predictor

The predictor for a non trivial guess at distance $\delta p$ from the bifurcation point is provided by the method

BifurcationKit.predictor — Method

predictor(
    nf::BifurcationKit.PeriodDoublingPO{<:Shooting},
    δp,
    ampfactor;
    override
) -> NamedTuple{(:orbitguess, :pnew, :prob, :ampfactor, :δp, :po), <:NTuple{6, Any}}

Compute the predictor for the period-doubling bifurcation of periodic orbit.

If override = true, then the predictor is simply x0 .+ ampfactor .* e for the parameter p0 + δp and where e is a bifurcating eigenvector.

Normal form based on Poincaré return map

Given a transversal section $\Sigma$ to $x_0$ at $x_0(0)$, the Poincaré return map $\mathcal P$ associates to each point $x\in\Sigma$ close to $x_0(0)$ the first point $\mathcal P(x,p)\in\Sigma$ where the orbit of (E) with initial condition $x$ intersects again $\Sigma$. Hence, the discrete map $x_{n+1}=\mathcal P(x_n,p)$ has normal form

\[x_{n+1} = (a_{11}\cdot(p-p_0)-1)\cdot x_n + cx_n^3 + ...\]

where [Kuz2]

\[c =\frac{1}{6}\left\langle \zeta^*, \mathcal{C}(\zeta, \zeta, \zeta)+3 \mathcal{B}\left(\zeta,\left(I_{n-1}-\mathcal{A}\right)^{-1} \mathcal{B}(\zeta, \zeta)\right)\right\rangle\]

where $\mathcal C=d_1^3\mathcal P(x_0(0), p_0)$, $\mathcal B = d_1^2\mathcal P(x_0(0), p_0)$ and $\mathcal A = d_1\mathcal P(x_0(0), p_0)$. Also:

\[\mathcal{A} \zeta=-\zeta, \mathcal{A}^{\mathrm{T}} \zeta^*=-\zeta^*\]

The coefficients a and c are returned in pd.nf.nf. The bifurcation is supercritical if $c > 0$ and subcritical if $c < 0$.

Large scale problems

The computation of the normal form is not optimized for Matrix-Free problems (e.g. Monodromy) yet.

Collocation case

The monodromy matrix and other flow differentials are computed using finite differences.

Normal form based on Iooss method

This is based on [Iooss],[Kuz2]. Suppose that the $T$ periodic orbit $x_0(\tau)$ has a Period-Doubling bifurcation for a parameter value $p_0$. Locally, the orbits can be represented by $p-p_0:=\mu$ and

\[x(\tau) = x_0(\tau)+\xi v(\tau)+H(\tau, \xi, \mu)\]

where

\[\left\{\begin{array}{l} \frac{d \tau}{d t}=1+a_{01}\cdot(p-p_0)+a_2 \xi^2+\cdots \\ \frac{d \xi}{d \tau}=c_{11}\cdot(p-p_0)\xi+c_3 \xi^3+\cdots \end{array}\right.\]

with center manifold correction $H(\tau, \xi, \mu)$ being $2T$ periodic in $\tau$ and $v(\tau)$ is a Floquet eigenvector for the eigenvalue -1.

The coefficients a₀₁, a₂, c₁₁ and c₃ are returned in pd.nf.nf. The bifurcation is supercritical if $c_3 < 0$ and subcritical if $c_3 > 0$.

See also

Detection of bifurcation points of periodic orbits

Continuation of Period-doubling (PD) bifurcations of periodic orbits

Period doubling in Lur'e problem (PD aBS): computation of the PD normal form (with both methods) and branch switching for all the periodic orbit methods

References

Kuznetsov

Yu. A. Kuznetsov, "Elements of Applied Bifurcation Theory", 2nd ed., 1998.

Kuz2

Kuznetsov et al., “Numerical Periodic Normalization for Codim 1 Bifurcations of Limit Cycles.”

Iooss

Iooss, "Global Characterization of the Normal Form for a Vector Field near a Closed Orbit.", 1988

predictor( nf::BifurcationKit.PeriodDoublingPO{<:Shooting}, δp, ampfactor; override ) -> NamedTuple{(:orbitguess, :pnew, :prob, :ampfactor, :δp, :po), <:NTuple{6, Any}}
Compute the predictor for the period-doubling bifurcation of periodic orbit.

predictor( gh::BifurcationKit.Bautin, ::Val{:FoldPeriodicOrbitCont}, ϵ; verbose, ampfactor ) -> NamedTuple{(:orbit, :ω, :params, :x0), <:Tuple{BifurcationKit.var"#786#789"{BifurcationKit.var"#FoldPO#788"{ϵ, q0, x0, h₂₀₀₀}} where {ϵ, q0, x0, h₂₀₀₀}, Any, Any, BifurcationKit.var"#787#790"}}
Compute the predictor for the curve of Folds of periodic orbits near the Bautin bifurcation point.

predictor( nf::BifurcationKit.BranchPointPO{<:Shooting}, δp, ampfactor; override ) -> NamedTuple{(:orbitguess, :pnew, :prob, :ampfactor, :po), <:Tuple{Any, Any, Shooting, Any, Any}}
Compute the predictor for the simple branch point of periodic orbit.

predictor( bt::BifurcationKit.BogdanovTakens, ::Val{:FoldCurve}, ds; verbose, ampfactor ) -> NamedTuple{(:fold, :EigenVec, :EigenVecAd, :x0), <:Tuple{BifurcationKit.var"#FoldCurve#741", BifurcationKit.var"#737#742"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#738#743"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#739#744"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#getx#740"{_A}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf, _A}}}
Compute the predictor for the Fold curve near the Bogdanov-Takens point.

predictor( bt::BifurcationKit.BogdanovTakens, ::Val{:HomoclinicCurve}, ds; verbose, ampfactor ) -> NamedTuple{(:α, :orbit), <:Tuple{BifurcationKit.var"#α#746", BifurcationKit.var"#xLP#752"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#u2#751"{a1, d, b, a}, BifurcationKit.var"#v1#750"{BifurcationKit.var"#v0#748", _A, _B}, BifurcationKit.var"#u1#749", BifurcationKit.var"#v0#748", BifurcationKit.var"#u0#747"} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf, a1, d, b, a, _A, _B}}}
Compute the predictor for the curve of homoclinic orbits near the Bogdanov-Takens point.

predictor( bt::BifurcationKit.BogdanovTakens, ::Val{:HopfCurve}, ds; verbose, ampfactor ) -> NamedTuple{(:hopf, :ω, :EigenVec, :EigenVecAd, :x0), <:Tuple{BifurcationKit.var"#726#733"{BifurcationKit.var"#HopfCurve#730"{par0, K2, K11, K10, b, a}} where {par0, K2, K11, K10, b, a}, BifurcationKit.var"#727#734"{BifurcationKit.var"#HopfCurve#730"{par0, K2, K11, K10, b, a}} where {par0, K2, K11, K10, b, a}, BifurcationKit.var"#EigenVec#731"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#getx#729"{_A}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf, _A}, BifurcationKit.var"#EigenVecAd#732"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#getx#729"{_A}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf, _A}, BifurcationKit.var"#728#735"{BifurcationKit.BogdanovTakens{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#getx#729"{_A}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf, _A}}}
Compute the predictor for the Hopf curve near the Bogdanov-Takens point.

predictor( hh::BifurcationKit.HopfHopf, ::Val{:HopfCurve}, ds; verbose, ampfactor ) -> NamedTuple{(:hopf, :ω, :EigenVec, :EigenVecAd, :x0), <:Tuple{BifurcationKit.var"#872#878"{BifurcationKit.var"#HopfCurve#875"{_A, _B}} where {_A, _B}, BifurcationKit.var"#873#879"{BifurcationKit.var"#HopfCurve#875"{_A, _B}} where {_A, _B}, BifurcationKit.var"#EigenVec#876"{BifurcationKit.HopfHopf{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#EigenVecAd#877"{BifurcationKit.HopfHopf{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#874#880"}}
Compute the predictor for the Hopf curve near the Hopf-Hopf bifurcation point.

predictor( hh::BifurcationKit.HopfHopf, ::Val{:NS}, ϵ; verbose, ampfactor ) -> NamedTuple{(:ns1, :ns2, :params1, :params2, :ω11, :ω12, :ω21, :ω22, :T1, :T2, :k1, :k2), <:Tuple{BifurcationKit.var"#882#886"{BifurcationKit.var"#NS1#884"{ϵ, q1, x1, h₂₀₀₀}} where {ϵ, q1, x1, h₂₀₀₀}, BifurcationKit.var"#883#887"{BifurcationKit.var"#NS2#885"{ϵ, q2, x2, h₀₀₂₀}} where {ϵ, q2, x2, h₀₀₂₀}, Vararg{Any, 10}}}
Compute the predictor for the curve of Neimark-Sacker points near the Hopf-Hopf bifurcation point.

struct Shooting{Tf<:BifurcationKit.AbstractFlow, Tjac<:BifurcationKit.AbstractJacobianType, Ts, Tsection, Tpar, Tlens} <: BifurcationKit.AbstractShootingDiscretization
Create a problem to implement the Simple / Parallel Multiple Standard Shooting method to locate periodic orbits / BVP.

predictor( zh::BifurcationKit.ZeroHopf, ::Val{:FoldCurve}, ds; verbose, ampfactor ) -> NamedTuple{(:fold, :λ0, :EigenVec, :EigenVecAd, :x0), <:Tuple{BifurcationKit.var"#833#839"{BifurcationKit.var"#FoldCurve#836"{_A, _B}} where {_A, _B}, BifurcationKit.var"#834#840"{BifurcationKit.var"#FoldCurve#836"{_A, _B}} where {_A, _B}, BifurcationKit.var"#EigenVec#837"{BifurcationKit.ZeroHopf{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#EigenVecAd#838"{BifurcationKit.ZeroHopf{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#835#841"}}
Compute the predictor for the curve of Fold bifurcations near the Zero-Hopf bifurcation point.

predictor( zh::BifurcationKit.ZeroHopf, ::Val{:HopfCurve}, ds; verbose, ampfactor ) -> NamedTuple{(:hopf, :ω, :EigenVec, :EigenVecAd, :x0), <:Tuple{BifurcationKit.var"#823#829"{BifurcationKit.var"#HopfCurve#826"{_A, _B}} where {_A, _B}, BifurcationKit.var"#824#830"{BifurcationKit.var"#HopfCurve#826"{_A, _B}} where {_A, _B}, BifurcationKit.var"#EigenVec#827"{BifurcationKit.ZeroHopf{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#EigenVecAd#828"{BifurcationKit.ZeroHopf{Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}} where {Tv, Tτ, T, Tpar, Tlens, Tevl, Tevr, Tnf}, BifurcationKit.var"#825#831"}}
Compute the predictor for the curve of Hopf bifurcations near the Zero-Hopf bifurcation point.

predictor( zh::BifurcationKit.ZeroHopf, ::Val{:NS}, ϵ; verbose, ampfactor ) -> NamedTuple{(:orbit, :hasNS, :params, :T, :k), <:Tuple{BifurcationKit.var"#843#845"{BifurcationKit.var"#NS#844"{_A, _B, _C}} where {_A, _B, _C}, Vararg{Any, 4}}}
Compute the predictor for the curve of Neimark-Sacker bifurcations near the Zero-Hopf bifurcation point.